requirements.txt
aioredis==1.3.1
alabaster==0.7.12
amqp==2.5.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
Babel==2.7.0
beautifulsoup4==4.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-admin-rangefilter==0.5.1
django-allauth==0.40.0
django-appconf==1.0.3
django-autoslug==1.9.6
django-bootstrap4==1.0.1
django-countries==5.5
django-crispy-forms==1.7.2
django-filter==2.2.0
django-paypal==1.0.0
djangorestframework==3.10.3
docutils==0.15.2
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
incremental==17.5.0
Jinja2==2.10.3
kombu==4.6.5
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.2
Naked==0.1.31
nose==1.3.7
oauthlib==3.1.0
packaging==19.2
paypalrestsdk==1.13.1
Pillow==6.2.1
pinax-referrals==3.0.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.9.0
Pygments==2.4.2
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pyparsing==2.4.2
python-decouple==3.1
python-mpesawrapper==1.0.4
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
service-identity==18.1.0
shellescape==3.4.1
six==1.12.0
snowballstemmer==2.0.0
soupsieve==1.9.4
Sphinx==2.2.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.6
vine==1.3.0
whitenoise==5.0.1
zipp==0.6.0
zope.interface==4.7.1
